// Angular Material ve Arayüz/dgsKitapcim/src/app/models/Siparis.ts
export class Siparis
{
  siparisId: number;
  siparisMusteriId: number;
  siparisKodu: number;
  siparisKitapId: number;
  siparisTarihi: string;
  siparisMusteriAdi: string;
  siparisMusteriSoyadi: string;
  siparisMusteriAdres: string;
  siparisMusteriTelefon: string;
  siparisMusteriMail: string;
  siparisKitapAdi: string;
  siparisKitapFiyati: number;
  siparisYayineviAdi: string;
  siparisKategoriAdi: string;
}